Add gallery posts and some cellular automata
I've decided to upgrade my website! The
<a href="/blog">blog</a> and <a href="/lab">lab</a>
are moving to the same page. I'm also going to add
more content types, like songs, galleries, and
videos. As I migrate things over, I'll be
backfilling my blog with content to test with.

The CA post is an example of filler content.